requirements.txt
streamlit
anthropic
feedparser
requests
python-dotenv
yfinance
